backend/agents/requirements.txt
google-adk
google-genai
google-cloud-aiplatform[adk,agent_engines]
